Test project was improved.

+++ DynamicMapper.Test/Domain/Person.cs
@@ -9,6 +9,6 @@ namespace DynamicMapper.Test.Domain
     {
         public string Name { get; set; }
         public string Surname { get; set; }
-        public decimal? AnnoNascita { get; set; }
+        public double? AnnoNascita { get; set; }
     }
 }
